+++ modulo4/projeto-labex/labex/src/routes/Routes.js
@@ -11,6 +11,7 @@ import ApplicationFormPage from '../pages/ApplicationFormPage';
 
 
 
+
 const Router = () => {
 
     return (
